+++ Source/Builders/Database/Restore-MockDatabase.ps1
@@ -57,7 +57,7 @@ $segments=@(
     $ishCDPath
     "Database"
     "Dump"
-    "SQLServer2014"
+    "SQLServer*"
 )
 
 if ($EmptyDB.IsPresent)